+++ adapters/postgres/postgres_test.go
@@ -7,10 +7,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/c2fo/testify/assert"
-	"github.com/c2fo/testify/suite"
+	"github.com/doug-martin/goqu"
+
 	"github.com/lib/pq"
-	"gopkg.in/doug-martin/goqu.v5"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/suite"
 )
 
 const schema = `
@@ -94,7 +95,7 @@ func (me *postgresTest) TestSelectSql() {
 
 	sql, args, err := ds.Prepared(true).Where(goqu.L("? = ?", goqu.I("int"), 10)).ToSql()
 	assert.NoError(t, err)
-	assert.Equal(t, args, []interface{}{10})
+	assert.Equal(t, args, []interface{}{int64(10)})
 	assert.Equal(t, sql, `SELECT * FROM "entry" WHERE "int" = $1`)
 }
 
@@ -202,19 +203,19 @@ func (me *postgresTest) TestCount() {
 	ds := me.db.From("entry")
 	count, err := ds.Count()
 	assert.NoError(t, err)
-	assert.Equal(t, count, 10)
+	assert.Equal(t, count, int64(10))
 	count, err = ds.Where(goqu.I("int").Gt(4)).Count()
 	assert.NoError(t, err)
-	assert.Equal(t, count, 5)
+	assert.Equal(t, count, int64(5))
 	count, err = ds.Where(goqu.I("int").Gte(4)).Count()
 	assert.NoError(t, err)
-	assert.Equal(t, count, 6)
+	assert.Equal(t, count, int64(6))
 	count, err = ds.Where(goqu.I("string").Like("0.1%")).Count()
 	assert.NoError(t, err)
-	assert.Equal(t, count, 1)
+	assert.Equal(t, count, int64(1))
 	count, err = ds.Where(goqu.I("string").IsNull()).Count()
 	assert.NoError(t, err)
-	assert.Equal(t, count, 0)
+	assert.Equal(t, count, int64(0))
 }
 
 func (me *postgresTest) TestInsert() {
@@ -303,7 +304,7 @@ func (me *postgresTest) TestUpdate() {
 
 	count, err := ds.Where(goqu.I("int").Eq(11)).Count()
 	assert.NoError(t, err)
-	assert.Equal(t, count, 1)
+	assert.Equal(t, count, int64(1))
 
 	var id uint32
 	found, err = ds.Where(goqu.I("int").Eq(11)).Returning("id").Update(map[string]interface{}{"int": 9}).ScanVal(&id)
@@ -324,7 +325,7 @@ func (me *postgresTest) TestDelete() {
 
 	count, err := ds.Count()
 	assert.NoError(t, err)
-	assert.Equal(t, count, 9)
+	assert.Equal(t, count, int64(9))
 
 	var id uint32
 	found, err = ds.Where(goqu.I("id").Eq(e.Id)).ScanVal(&id)
@@ -335,7 +336,7 @@ func (me *postgresTest) TestDelete() {
 	found, err = ds.Where(goqu.I("int").Eq(8)).Select("id").ScanStruct(&e)
 	assert.NoError(t, err)
 	assert.True(t, found)
-	assert.NotEqual(t, e.Id, 0)
+	assert.NotEqual(t, e.Id, int64(0))
 
 	id = 0
 	_, err = ds.Where(goqu.I("id").Eq(e.Id)).Returning("id").Delete().ScanVal(&id)
@@ -359,7 +360,7 @@ func (me *postgresTest) TestInsertIgnore() {
 
 	count, err := ds.Count()
 	assert.NoError(t, err)
-	assert.Equal(t, count, 11)
+	assert.Equal(t, count, int64(11))
 }
 
 func (me *postgresTest) TestInsertConflict() {